+++ src/kinetic_models/ButlerVolmer.jl
@@ -11,7 +11,7 @@ struct ButlerVolmer{T} <: NonIntegralModel{T}
     α::T
     function ButlerVolmer(A, α)
         @assert all(α .<= 1.0 .&& α .>=0.0) "Electron transfer coefficient must be in [0,1]"
-        ps = consistent_params(Float64.(A), Float64.(α))
+        ps = consistent_params(Float32.(A), Float32.(α))
         new{typeof(ps[1])}(ps...)
     end
 end
